@@ -6,58 +6,58 @@
     "start": "ng serve --proxy-config proxy.conf.json",
     "build": "ng build",
     "build-prod": "ng build --optimization=true --prod=true --sourceMap=false",
-    "transpile-shared": "babel ../shared/capacity.js --out-file ../shared/capacity.es5.js",
+    "transpile-shared": "babel ../shared/capacity.js --out-file src/shared/capacity.es5.js",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.1",
-    "@angular/cdk": "^7.2.1",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "~7.2.0",
-    "@angular/material": "^7.2.1",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "~8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "^8.0.0",
+    "@angular/material": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "@types/moment": "^2.13.0",
     "chart.js": "^2.8.0",
-    "core-js": "^2.6.9",
+    "core-js": "^3.1.3",
     "hammerjs": "^2.0.8",
     "hash.js": "^1.1.7",
     "moment": "^2.24.0",
-    "rxjs": "~6.5.2",
+    "rxjs": "^6.5.2",
     "rxjs-compat": "^6.5.2",
     "socket.io-client": "^2.2.0",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.29"
+    "tslib": "^1.9.3",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.9",
-    "@angular/cli": "~7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "~7.2.15",
+    "@angular-devkit/build-angular": "^0.800.0",
+    "@angular/cli": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
     "@babel/cli": "^7.4.4",
     "@babel/core": "^7.4.5",
     "@babel/preset-env": "^7.4.5",
-    "@types/jasmine": "~3.3.13",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~12.0.3",
-    "codelyzer": "~5.0.1",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.5",
-    "karma-jasmine": "~2.0.1",
+    "@types/jasmine": "^3.3.13",
+    "@types/jasminewd2": "^2.0.6",
+    "@types/node": "^12.0.3",
+    "codelyzer": "^5.0.1",
+    "jasmine-core": "^3.4.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^4.1.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-jasmine": "^2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "~6.0.0",
-    "ts-node": "~8.2.0",
-    "tslint": "~5.16.0",
-    "typescript": "~3.2.4"
+    "ts-node": "^8.2.0",
+    "tslint": "^5.16.0",
+    "typescript": "^3.4.5"
   }
 }
