@@ -10,7 +10,7 @@ const presets = [
         ie: "11",
       },
       useBuiltIns: "usage",
-      corejs: '2.6.5',
+      corejs: '3.1.3',
     },
   ],
 ];
